+++ company-box.el
@@ -587,7 +587,7 @@ It doesn't nothing if a font icon is used."
   (company-box--compute-frame-position (company-box--get-frame))
   (company-box--move-selection t)
   (company-box--update-frame-position (company-box--get-frame))
-  (company-box--start-frame-timer t)
+  (company-box--show-frame t)
   (company-box--update-scrollbar (company-box--get-frame) t)
   (company-box--with-buffer nil
     (company-box--maybe-move-number (or company-box--last-start 1))))
@@ -758,27 +758,11 @@ It doesn't nothing if a font icon is used."
 
 (defvar company-box-hide-hook nil)
 
-(defvar company-box--frame-show-timer nil)
+(defvar company-box--frame-timer nil)
 
-(defun company-box--frame-show (show)
-  "If SHOW is non-nil, make the frame visible; otherwise make it invisible."
-  (if-let* ((local-frame (company-box--get-frame))
-            ((frame-live-p local-frame)))
-      (let ((visible (frame-visible-p local-frame))
-            (func (if show #'make-frame-visible #'make-frame-invisible)))
-        (unless (eq show visible)
-          (funcall func local-frame)))
-    (unless local-frame
-      (company-box--set-frame (company-box--make-frame)))
-    (company-box--start-frame-timer show)))
-
-(defun company-box--start-frame-timer (show)
-  "Delay show/hide frame to prevent GUI bug.
-
-Argument SHOW, see function `company-box--frame-show' description."
-  (company-box--kill-timer company-box--frame-show-timer)
-  (setq company-box--frame-show-timer
-        (run-with-idle-timer 0 nil #'company-box--frame-show show)))
+(defun company-box--show-frame (show)
+  "Start the timer to SHOW frame."
+  (company-box--start-frame-timer show (company-box--get-frame) 'company-box--frame-timer))
 
 (defun company-box-hide ()
   "Hide the completion window."
@@ -787,7 +771,7 @@ Argument SHOW, see function `company-box--frame-show' description."
         company-box--prefix-pos nil
         company-box--last-start nil
         company-box--edges nil)
-  (company-box--start-frame-timer nil)
+  (company-box--show-frame nil)
   (company-box--with-buffer nil
     (setq company-box--last-start nil))
   (remove-hook 'window-scroll-functions 'company-box--handle-scroll-parent t)